Add RequestIdMiddleware for tracking HTTP request IDs
Introduced a new namespace `AdventureWorks.Middlewares.RequestId`
and added the `RequestIdMiddleware` class to implement HTTP
middleware functionality. The middleware generates a unique
`requestId` for each request, stores it in `HttpContext.Items`,
and adds it to the response headers using `Constants.RequestId`.
The `InvokeAsync` method ensures the middleware integrates
seamlessly with the pipeline by invoking the next delegate.